Gjort input-övningar

+++ ovningar/variabler/ovning10.py
@@ -5,7 +5,7 @@ dollar = sek = 0
 dollar = input("Hur många dollar vill du omvandla till SEK?")
 
 # För att omvandla dollar till SEK
-sek = int(dollar) * dollarkurs
+sek = float(dollar) * dollarkurs
 
 # Visar resultatet av omvandlingen för användaren
 print(str(dollar) + " dollar motsvarar " + str(sek) + " svenska kronor")
